+++ src/client/online/moderator/UserCollections.tsx
@@ -1,5 +1,5 @@
-import { useInfiniteQuery } from '@tanstack/react-query';
-import { memo, useMemo } from 'react';
+import { useInfiniteQuery, useMutation } from '@tanstack/react-query';
+import { memo, RefObject, useMemo } from 'react';
 import Loading from '../../components/Loading';
 import InfiniteScrollTrigger from '../../components/InfiniteScrollTrigger';
 import { CollectionBrief, ResourceStatus } from '../data';
@@ -10,28 +10,130 @@ import { searchCollectionsInfiniteQueryOptions } from '../CollectionSearchResult
 import { TbLayoutGrid } from 'react-icons/tb';
 import Skeleton from '../../components/Skeleton';
 import { Link } from '@tanstack/react-router';
+import { modPrompt, PromptHandle } from './ModMessagePrompt';
+import { BsThreeDotsVertical } from 'react-icons/bs';
+import { api, queryClient } from '../api';
+import toast from 'react-hot-toast';
 
 const UserCollection = memo(function UserCollection({
   collection,
+  promptHandle,
 }: {
   collection: CollectionBrief;
+  promptHandle: RefObject<PromptHandle | null>;
 }) {
+  const updateCollection = useMutation({
+    mutationFn: (data: Parameters<typeof api.modUpdateCollection>) => {
+      return api.modUpdateCollection(...data);
+    },
+    onError(error) {
+      toast.error(error.message);
+    },
+    onSuccess() {
+      void queryClient.invalidateQueries({
+        queryKey: ['collection', 'search'],
+      });
+    },
+  });
+  const removeDescription = useMutation({
+    mutationFn: (
+      data: Parameters<typeof api.modRemoveCollectionDescription>
+    ) => {
+      return api.modRemoveCollectionDescription(...data);
+    },
+    onError(error) {
+      toast.error(error.message);
+    },
+    onSuccess() {
+      void queryClient.invalidateQueries({
+        queryKey: ['collection', 'search'],
+      });
+    },
+  });
   return (
     <div className="flex flex-col items-start self-stretch gap-1 shrink-0">
-      <Link
-        to="/collection/$collectionId"
-        params={{ collectionId: collection.id }}
-        className="text-lg"
-      >
-        {collection.isSeries && (
-          <FaListOl size={14} className="inline text-accent" />
-        )}{' '}
-        {collection.title.length === 0 ? (
-          <span className="opacity-80">Untitled Collection</span>
-        ) : (
-          collection.title
-        )}
-      </Link>
+      <div className="w-full flex justify-between items-center gap-2">
+        <Link
+          to="/collection/$collectionId"
+          params={{ collectionId: collection.id }}
+          className="text-lg"
+        >
+          {collection.isSeries && (
+            <FaListOl size={14} className="inline text-accent" />
+          )}{' '}
+          {collection.title.length === 0 ? (
+            <span className="opacity-80">Untitled Collection</span>
+          ) : (
+            collection.title
+          )}
+        </Link>
+        <div className="dropdown dropdown-end">
+          <div tabIndex={0} role="button" className="btn btn-xs btn-ghost m-1">
+            <BsThreeDotsVertical size={16} />
+          </div>
+          <ul
+            tabIndex={-1}
+            className="dropdown-content menu bg-base-100 rounded-box z-1 w-52 p-2 shadow-sm"
+          >
+            <li>
+              <a
+                onClick={() =>
+                  modPrompt(
+                    promptHandle,
+                    'Removing collection description. Please explain why this action is being taken. This message will be sent to the user and logged with the action.'
+                  )
+                    .then(message =>
+                      removeDescription.mutate([collection.id, message])
+                    )
+                    .catch(() => {})
+                }
+              >
+                Remove description
+              </a>
+            </li>
+            <li>
+              <a
+                onClick={() =>
+                  modPrompt(
+                    promptHandle,
+                    'Unpublishing collection. Please explain why this action is being taken. This message will be sent to the user and logged with the action.'
+                  )
+                    .then(message =>
+                      updateCollection.mutate([
+                        'unpublish',
+                        collection.id,
+                        message,
+                      ])
+                    )
+                    .catch(() => {})
+                }
+              >
+                Unpublish
+              </a>
+            </li>
+            <li>
+              <a
+                onClick={() =>
+                  modPrompt(
+                    promptHandle,
+                    'Deleting collection. Please explain why this action is being taken. This message will be sent to the user and logged with the action.'
+                  )
+                    .then(message =>
+                      updateCollection.mutate([
+                        'delete',
+                        collection.id,
+                        message,
+                      ])
+                    )
+                    .catch(() => {})
+                }
+              >
+                Delete
+              </a>
+            </li>
+          </ul>
+        </div>
+      </div>
       <div className="text-xs opacity-80 flex flex-wrap gap-2">
         <span>Created {toRelativeDate(new Date(collection.createdAt))}</span>
         <span>Updated {toRelativeDate(new Date(collection.updatedAt))}</span>
@@ -61,9 +163,13 @@ const UserCollection = memo(function UserCollection({
 
 export interface UserCollectionsProps {
   userId: string;
+  promptHandle: RefObject<PromptHandle | null>;
 }
 
-export default memo(function UserCollections({ userId }: UserCollectionsProps) {
+export default memo(function UserCollections({
+  userId,
+  promptHandle,
+}: UserCollectionsProps) {
   const searchParams = useMemo<CollectionSearchParams>(
     () => ({
       q: `creator=${userId}`,
@@ -99,7 +205,11 @@ export default memo(function UserCollections({ userId }: UserCollectionsProps) {
           <div className="flex flex-col gap-4 items-center">
             {data?.pages.flatMap(page =>
               page.results.map(collection => (
-                <UserCollection key={collection.id} collection={collection} />
+                <UserCollection
+                  key={collection.id}
+                  collection={collection}
+                  promptHandle={promptHandle}
+                />
               ))
             )}
             {isFetching ? (